create class for Transaction methods
